domain/requirements.txt
fastapi
neo4j
dnspython
pytest
sqlalchemy